Se revisan migraciones, tiene errores

+++ src/database/migrations/2017_09_18_201139_create_plan_de_pagos_table.php
@@ -17,10 +17,11 @@ class CreatePlanDePagosTable extends Migration
             $table->increments('id_plan_de_pago');
             $table->string('nombre_plan');
             $table->integer('cuotas');
-            $table->float('valor_cuota');
-            $table->float('interes');
+            $table->decimal('valor_cuota', 8, 2);
+            $table->decimal('interes', 8, 2);
             $table->string('forma_pago');
             $table->timestamps();
+
         });
     }
